refactor: Introduce sub-commands for the verify command
This commit restructures the verify command to include a 'reject' sub-command.
The current verify command supports listening to the reject messages and is therefore moved to the reject sub-command.
To address issue #3, this refactoring is meant to
make ready for the inclusion of a 'confirm' sub-command.
