+++ internal/verify/reject/reject.go
@@ -1,4 +1,4 @@
-package verifyImplementation
+package rejectImplementation
 
 import (
 	"context"
