+++ internal/verify/reject/reject_test.go
@@ -1,4 +1,4 @@
-package verifyImplementation
+package rejectImplementation
 
 import (
 	"testing"
