+++ internal/verify/reject/teams_message.go
@@ -1,4 +1,4 @@
-package verifyImplementation
+package rejectImplementation
 
 import (
 	"fmt"
