+++ internal/verify/reject/teams_message_test.go
@@ -1,4 +1,4 @@
-package verifyImplementation
+package rejectImplementation
 
 import (
 	"encoding/json"
